+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVRY2FBYJJ8JH77BDD0BT0XE"
+        "01JVS87NPM0HQF90FYB97WKGD9"
     )
     r"""The id of the resource"""