+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-21T08:06:28.385Z")
+    created: Optional[datetime] = parse_datetime("2025-05-21T11:04:04.625Z")
     r"""The date and time the resource was created"""