+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVRY2FBXMBWV5YAT8CRY8XEV"
+        "01JVS87NPKK390MBGS6KMCQNBA"
     )
     r"""The id of the resource"""