+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -223,5 +223,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-21T08:06:25.552Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-21T11:04:01.937Z")
     r"""The date and time the resource was last updated"""