+++ packages/orq-rc/RELEASES.md
@@ -1359,3 +1359,13 @@ Based on:
 - [python v3.6.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.21] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.21 - packages/orq-rc
+
+## 2025-05-21 11:05:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.3 (2.605.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.22 - packages/orq-rc